+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.0.4
+### Added
+- New `conflict` event on Node instances (only happens when a conflict overwrites the current value).
+
+### Changed
+- `Node#compare` returns `"conflict"` on conflicts, instead of attempting to interpret them as `"update"` or `"history"`.
+
 ## v0.0.3
 ### Changed
 - Node `historical` event renamed to `history`.